added image and category to product/changed from postgres to mysql due to familiarity

+++ src/main/java/lemos/bruno/myerp/domain/dtos/ProductDto.java
@@ -1,6 +1,8 @@
 package lemos.bruno.myerp.domain.dtos;
 
-public record ProductDto(String name, String description, Double price, Double stock) {
+import java.math.BigDecimal;
+
+public record ProductDto(String name, String description, BigDecimal price, Double stock) {
 
 
 }
